Rename blocks.js plugin to inert-blocks
