+++ app/app.js
@@ -36,7 +36,7 @@ var main = function() {
     'voxel-wool': require('voxel-wool'),
     'voxel-pumpkin': require('voxel-pumpkin'),
     'voxel-blockdata': require('voxel-blockdata'),
-    './blocks.js': require('./blocks.js'),
+    './inert-blocks': require('./inert-blocks'),
     './inert-items': require('./inert-items'),
     'voxel-decorative': require('voxel-decorative'),
     'voxel-land': require('voxel-land'),
@@ -120,7 +120,7 @@ var main = function() {
     'voxel-wool': {},
     'voxel-pumpkin': {},
 
-    './blocks.js': {}, // misc inanimate opaque solid blocks
+    './inert-blocks': {}, // misc inanimate opaque solid blocks
     './inert-items': {},
     'voxel-decorative': {},
     //'voxel-land': {registerBlocks: false},
